backend/requirements.txt
# Web 框架
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# 数据库
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# 缓存
redis[hiredis]==5.0.4

# Elasticsearch
elasticsearch[async]==8.13.0

# 任务队列
celery==5.4.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 配置
python-dotenv==1.0.1
pydantic-settings==2.2.1

# MinIO 文件存储
minio==7.2.7

# 工具库
httpx==0.27.0
aiofiles==23.2.1
openpyxl==3.1.2

# 测试
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# 代码质量
black==24.4.2
flake8==7.0.0
